My source code, training and test dataset
I developed a source retrieval model which addresses the scalability in two level: Rough
Search and Fine Search. The implementation of Rough Search is based on 1-gram search methode
similar to public search engines such as google and yahoo which is highley scalable and mainly
depends on the off-line pre-processing. I increased the accuracy of the Rough Search by using text
chunking and boosting.
The implementation of Fine Search is based in N-gram feature extraction which should be done in
real-time. Since the Fine Search uses data chunking, it can be easily implemented in Map-Reduce
framework over each chunck of data.
The Rough Search stochastically provides a feasible search space and Fine Search deterministaclly
finds whether the observation is from a source in the database or not.
Through my model evaluation, I tried to understand the sensitivity of my model based on different
features. More specifically, the size of N-gram feature play very important